changed divideWork in mpiController to return vector<long> instead of <int>

+++ src/bands/bandstructure.h
@@ -82,7 +82,7 @@ class BaseBandStructure {
   /** Returns an iterator to be used for loops over the Bloch state index.
    * The values of the iterator are distributed in N blocks over N MPI ranks.
    */
-  std::vector<int> parallelStateIterator();
+  std::vector<long> parallelStateIterator();
 
   /** Returns the energy of a quasiparticle from its Bloch index
    * Used for accessing the bandstructure in the BTE.
