+++ src/mpi/mpiController.h
@@ -135,11 +135,11 @@ class MPIcontroller {
    * @return divs: returns a vector of length 2, containing start and stop
    *       points for the divided number of tasks.
    */
-  std::vector<int> divideWork(size_t numTasks);  // divide up a set of work
+  std::vector<long> divideWork(size_t numTasks);  // divide up a set of work
   /** Divides a number of tasks appropriately for the current MPI env.
    * @return divs: returns an iterator of points for the divided number of tasks.
    */
-  std::vector<int> divideWorkIter(size_t numTasks);
+  std::vector<long> divideWorkIter(size_t numTasks);
 
 };
